starter_ai_agents/ai_personal_assistant_agent/requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-google-genai>=2.0.0
python-dotenv>=1.0.1
requests>=2.32.0
ddgs>=9.0.0
